+++ packages/pluggableWidgets/combobox-web/src/Combobox.editorPreview.tsx
@@ -3,7 +3,11 @@ import { ReferenceValue } from "mendix";
 import { ComponentType, ReactElement, ReactNode, createElement, useMemo } from "react";
 import { ComboboxPreviewProps, OptionsSourceAssociationCustomContentTypeEnum } from "../typings/ComboboxProps";
 import { SingleSelection } from "./components/SingleSelection/SingleSelection";
-import { AssociationSimpleCaptionsProvider } from "./helpers/Association/AssociationSimpleCaptionsProvider";
+import {
+    AssociationSimpleCaptionsProvider,
+    CaptionContent
+} from "./helpers/Association/AssociationSimpleCaptionsProvider";
+import { AssociationOptionsProvider } from "./helpers/Association/AssociationOptionsProvider";
 import { BaseAssociationSelector } from "./helpers/Association/BaseAssociationSelector";
 import { CaptionPlacement, SingleSelector } from "./helpers/types";
 import { getDatasourcePlaceholderText } from "./helpers/utils";
@@ -12,6 +16,12 @@ import "./ui/Combobox.scss";
 interface PreviewProps {
     customContentRenderer: ComponentType<{ children: ReactNode; caption?: string }>;
     customContentType: OptionsSourceAssociationCustomContentTypeEnum;
+}
+
+class AssociationPreviewOptionsProvider extends AssociationOptionsProvider {
+    getAll(): string[] {
+        return ["..."];
+    }
 }
 
 class AssociationPreviewCaptionsProvider extends AssociationSimpleCaptionsProvider {
@@ -39,16 +49,23 @@ class AssociationPreviewCaptionsProvider extends AssociationSimpleCaptionsProvid
         this.customContentType = props.customContentType;
     }
 
-    render(value: string | null, placement: CaptionPlacement): ReactNode {
+    render(value: string | null, placement: CaptionPlacement, htmlFor?: string): ReactNode {
         // always render custom content dropzone in design mode if type is options only
+        if (placement === "options") {
+            return <CaptionContent htmlFor={htmlFor}>{this.get(value)}</CaptionContent>;
+        }
+
         return super.render(value, placement === "label" ? "options" : placement);
+        // return super.render(value, placement);
     }
 }
 class AssociationPreviewSelector extends BaseAssociationSelector<string, ReferenceValue> implements SingleSelector {
     type = "single" as const;
+
     constructor(props: ComboboxPreviewProps) {
         super();
         this.caption = new AssociationPreviewCaptionsProvider(new Map());
+        this.options = new AssociationPreviewOptionsProvider(this.caption, new Map());
         this.readOnly = props.readOnly;
         this.clearable = props.clearable;
         this.currentValue = getDatasourcePlaceholderText(props);
@@ -83,7 +100,14 @@ export const preview = (props: ComboboxPreviewProps): ReactElement => {
                 a11yInstructions: props.a11yInstructions,
                 a11yNoOption: props.noOptionsText
             }
-        }
+        },
+        showFooter: props.showFooter,
+        menuFooterContent: (
+            <props.menuFooterContent.renderer caption="Place footer widget here">
+                <div />
+            </props.menuFooterContent.renderer>
+        ),
+        keepMenuOpen: props.showFooter
     };
 
     // eslint-disable-next-line react-hooks/rules-of-hooks