+++ packages/pluggableWidgets/combobox-web/src/components/SingleSelection/SingleSelectionMenu.tsx
@@ -1,5 +1,5 @@
 import { UseComboboxPropGetters } from "downshift/typings";
-import { createElement, ReactElement } from "react";
+import { createElement, ReactElement, ReactNode } from "react";
 import { SingleSelector } from "../../helpers/types";
 import { ComboboxMenuWrapper } from "../ComboboxMenuWrapper";
 import { ComboboxOptionWrapper } from "../ComboboxOptionWrapper";
@@ -10,6 +10,9 @@ interface ComboboxMenuProps extends Partial<UseComboboxPropGetters<string>> {
     highlightedIndex: number | null;
     selectedItem?: string | null;
     noOptionsText?: string;
+    alwaysOpen?: boolean;
+    showFooter: boolean;
+    menuFooterContent?: ReactNode;
 }
 
 export function SingleSelectionMenu({
@@ -18,21 +21,28 @@ export function SingleSelectionMenu({
     highlightedIndex,
     getMenuProps,
     getItemProps,
-    noOptionsText
+    noOptionsText,
+    alwaysOpen,
+    showFooter,
+    menuFooterContent
 }: ComboboxMenuProps): ReactElement {
     const items = selector.options.getAll();
+
     return (
         <ComboboxMenuWrapper
             isOpen={isOpen}
             isEmpty={items?.length <= 0}
             getMenuProps={getMenuProps}
             noOptionsText={noOptionsText}
+            showFooter={showFooter}
+            menuFooterContent={menuFooterContent}
+            alwaysOpen={alwaysOpen}
         >
             {isOpen &&
                 items.map((item, index) => (
                     <ComboboxOptionWrapper
                         key={item}
-                        isHighlighted={highlightedIndex === index}
+                        isHighlighted={alwaysOpen ? false : highlightedIndex === index}
                         isSelected={selector.currentValue === item}
                         item={item}
                         getItemProps={getItemProps}